feat: split map view to its  own component, added a base styling sheet.  render map in main.jsx
