+++ src/MapView.css
@@ -1,0 +1,14 @@
+.container-view {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    width: 100%;
+  }
+  
+  .map-view {
+    display: flex;
+    margin-left: 10%;
+
+  }
